Extract BulkOperation class
